Refactor GraphQL error handling to one single class
